// packages/joplin-api/src/model/NoteUpdateRes.ts
import { NoteProperties } from './NoteProperties'
import { CommonType } from './CommonType'

export type NoteUpdateRes = Pick<
  NoteProperties,
  | 'id'
  | 'parent_id'
  | 'title'
  | 'body'
  | 'created_time'
  | 'updated_time'
  | 'is_conflict'
  | 'latitude'
  | 'longitude'
  | 'altitude'
  | 'author'
  | 'source_url'
  | 'is_todo'
  | 'todo_due'
  | 'todo_completed'
  | 'source'
  | 'source_application'
  | 'application_data'
  | 'order'
  | 'user_created_time'
  | 'user_updated_time'
  | 'encryption_cipher_text'
  | 'encryption_applied'
  | 'markup_language'
  | 'is_shared'
> &
  CommonType